+++ nilearn/datasets/tests/test_atlas.py
@@ -321,32 +321,66 @@ def test_fetch_coords_seitzman_2018(request_mocker):
     assert np.any(bunch.networks != np.sort(bunch.networks))
 
 
-def _destrieux_data():
+def _destrieux_data(n_labels=10):
+    """Function mocking the download of the destrieux atlas.
+    Parameter `n_labels` controls the number of labels in the
+    CSV file. The number of labels in the image is 10.
+    """
     data = {"destrieux2009.rst": "readme"}
+    atlas = np.random.randint(0, 10, (10, 10, 10))
+    atlas_img = nibabel.Nifti1Image(atlas, np.eye(4))
+    labels = "\n".join([f"{idx},label {idx}" for idx in range(n_labels)])
+    labels = "index,name\n" + labels
     for lat in ["_lateralized", ""]:
         lat_data = {
-            "destrieux2009_rois_labels{}.csv".format(lat): "name,index",
-            "destrieux2009_rois{}.nii.gz".format(lat): "",
+            "destrieux2009_rois_labels{}.csv".format(lat): labels,
+            "destrieux2009_rois{}.nii.gz".format(lat): atlas_img,
         }
         data.update(lat_data)
     return dict_to_archive(data)
 
 
-def test_fetch_atlas_destrieux_2009(tmp_path, request_mocker):
-    request_mocker.url_mapping["*destrieux2009.tgz"] = _destrieux_data()
-    bunch = atlas.fetch_atlas_destrieux_2009(data_dir=tmp_path,
-                                             verbose=0)
-
-    assert request_mocker.url_count == 1
-    assert bunch['maps'] == str(tmp_path / 'destrieux_2009'
-                                / 'destrieux2009_rois_lateralized.nii.gz')
-
+@pytest.mark.parametrize("n_labels", [10, 11])
+@pytest.mark.parametrize("lateralized", [True, False])
+@pytest.mark.parametrize("clean_labels", [True, False])
+def test_fetch_atlas_destrieux_2009(tmp_path, request_mocker, n_labels,
+                                    lateralized, clean_labels):
+    """Tests for function `fetch_atlas_destrieux_2009`.
+    The atlas is fetched with different combinations of `lateralized`
+    and `n_labels`. In the case `n_labels=11`, the fetcher should be able
+    to remove the extra label ('10') not present in the atlas image.
+    """
+    request_mocker.url_mapping["*destrieux2009.tgz"] = _destrieux_data(
+        n_labels=n_labels
+    )
     bunch = atlas.fetch_atlas_destrieux_2009(
-        lateralized=False, data_dir=tmp_path, verbose=0)
-
+        lateralized=lateralized, data_dir=tmp_path,
+        verbose=0, clean_labels=clean_labels
+    )
     assert request_mocker.url_count == 1
-    assert bunch['maps'] == str(tmp_path / 'destrieux_2009'
-                                / 'destrieux2009_rois.nii.gz')
+    name = '_lateralized' if lateralized else ""
+    assert bunch['maps'] == str(
+        tmp_path / 'destrieux_2009' / f'destrieux2009_rois{name}.nii.gz'
+    )
+    labels_img = set(np.unique(get_data(bunch.maps)))
+    labels = set([label.index for label in bunch.labels])
+    if clean_labels:
+        assert labels_img == labels
+    else:
+        assert labels_img.issubset(labels)
+
+
+def test_fetch_atlas_destrieux_2009_warning_clean_labels(request_mocker):
+    """Tests that ``fetch_atlas_destrieux_2009`` gives a FutureWarning
+    when ``clean_labels`` isn't specified explicitly.
+    """
+    request_mocker.url_mapping["*destrieux2009.tgz"] = _destrieux_data()
+    with pytest.warns(FutureWarning,
+                      match=("Default value for parameter `clean_labels` "
+                             "will change from ``False`` to ``True`` "
+                             "in Nilearn 0.10.")):
+        atlas.fetch_atlas_destrieux_2009()
+    assert request_mocker.url_count == 1
 
 
 def test_fetch_atlas_msdl(tmp_path, request_mocker):